1. Remove MvcPaging using in Web.Config. 2. Add AssertConfiguratuinIsValid in AutoMapperConfig. 3. Modify RecordViewModelMapToEntityProfile. 4. Chage return type from IEnumerable to IQueryable at getAll( ) in RecordService. 5. Add AutoMapper.QueryableExtesions in RecordController.

+++ MyBookKeeping/AutoMapper/RecordViewModelMapToEntityProfile.cs
@@ -11,7 +11,7 @@ namespace MyBookKeeping.AutoMapper
         public RecordViewModelMapToEntityProfile( )
         {
             CreateMap<AccountBook, RecordViewModel>( )
-                .ForMember( x => x.Amount, y => y.MapFrom( s => s.Amounttt ) )
+                .ForMember( x => x.Amount, y => y.MapFrom( s => ( decimal ) s.Amounttt ) )
                 .ForMember( x => x.Category, y => y.MapFrom( s => s.Categoryyy == 0 ? "收入" : "支出" ) )
                 .ForMember( x => x.Date, y => y.MapFrom( s => s.Dateee ) );
         }
